Shakedown. Updated return, started description
Datetime coersion, location coercion and coordinate generation, class description, better naming
!!!__UNTESTED__!!!
